+++ portality/formcontext/xwalks/journal_form.py
@@ -1,7 +1,7 @@
 from portality import models, lcc
 from portality.datasets import licenses
 from portality.formcontext.choices import Choices
-from portality.formcontext.xwalks.interprets import interpret_special, interpret_other, reverse_interpret_special, \
+from portality.formcontext.xwalks.interpreting_methods import interpret_special, interpret_other, reverse_interpret_special, \
     reverse_interpret_other
 from portality.formcontext.xwalks.journal_generic import JournalGenericXWalk
 from portality.util import listpop
